order_executor/requirements.txt
redis
grpcio==1.70.0
grpcio-tools==1.70.0
protobuf==5.29.3
watchdog==6.0.0
flask
